MNTR: creating a router instead of app.get() etc.
