+++ task2/src/index.ts
@@ -1,85 +1,14 @@
 import * as dotenv from 'dotenv';
 dotenv.config();
-import express, { Request, Response } from 'express';
-import { ValidatedRequest } from 'express-joi-validation';
-import { v4 as uuid } from 'uuid';
-import { User } from './types/User';
-import { getAutoSuggestUsers } from './utils/getAutoSuggestUsers';
-import { UserRequestSchema, validator } from './middleware/validator';
-import { UserValidationSchema } from './schemas/User';
+import express from 'express';
+import { userRouter } from './router/userRouter';
 
 const app = express();
-
 const PORT = process.env.API_PORT || 3000;
-
-export let users: User[] = [
-    { age: 23, isDeleted: false, login: 'Dmytro', password: 'admin', id: uuid() },
-    { age: 24, isDeleted: false, login: 'Pavlo', password: 'mod', id: uuid() },
-    { age: 27, isDeleted: false, login: 'Oleh', password: 'guest', id: uuid() },
-    { age: 18, isDeleted: false, login: 'Zekora', password: 'guest', id: uuid() },
-    { age: 44, isDeleted: false, login: 'Artem', password: 'guest', id: uuid() },
-    { age: 20, isDeleted: false, login: 'Daniel', password: 'mod', id: uuid() }
-];
 
 app.use(express.json());
 
-app.param('id', (req: Request, res: Response, next, id) => {
-    const userById = users.find((user) => user.id === id);
-    if (userById) req.user = userById;
-    next();
-});
-
-app.get('/users', (req: Request, res: Response) => {
-    let { limit, filter } = req.query;
-    if (!limit) limit = '10';
-    if (!filter) filter = '';
-    const suggestedUsers = getAutoSuggestUsers(String(filter), String(limit));
-
-    res.json(suggestedUsers);
-});
-
-app.get('/user/:id', (req: Request, res: Response) => {
-    const userById = req.user;
-    res.json(userById);
-});
-
-app.post('/user', validator.body(UserValidationSchema),
-    (req: ValidatedRequest<UserRequestSchema>, res: Response) => {
-        const user: User = { ...req.body, id: uuid() };
-        users.push(user);
-        res.status(201);
-        res.send(`User with id ${user.id} successfully created`);
-    });
-
-app.put('/user/:id', validator.body(UserValidationSchema),
-    (req: ValidatedRequest<UserRequestSchema>, res: Response) => {
-        const { id } = req.user;
-        const userFromBody = req.body;
-        users = users.map((user: User) => {
-            if (user.id === id) {
-                return { ...userFromBody, id };
-            }
-
-            return user;
-        });
-        res.json(`User with id ${id} successfully edited`);
-    });
-
-app.delete('/user/:id', (req: Request, res: Response) => {
-    const { id } = req.user;
-    users = users.map((user: User) => {
-        if (user.id === id) {
-            return { ...user, isDeleted: true };
-        }
-
-        return user;
-    });
-    res.json(`User with id ${id} successfully deleted`);
-});
-
-app.use((req, res) => {
-    res.status(404).json({ message: 'Path not found' });
-});
+app.use('/', userRouter);
 
 app.listen(PORT, () => {
     console.log(`Server was started at port ${PORT}`);
